TBB-63 rename: Rename data files
