--- android/app/src/main/java/com/littlekakao/travelblahblah/data/local/db/dao/JourneyDao.kt
@@ -1,4 +1,0 @@
-package com.littlekakao.travelblahblah.data.local.db.dao
-
-interface JourneyDao {
-}
