--- android/app/src/main/java/com/littlekakao/travelblahblah/data/local/db/dao/UserDao.kt
@@ -1,4 +1,0 @@
-package com.littlekakao.travelblahblah.data.local.db.dao
-
-interface UserDao {
-}
