+++ android/app/src/main/java/com/littlekakao/travelblahblah/data/model/api/login/LoginRequest.kt
@@ -1,4 +1,3 @@
 package com.littlekakao.travelblahblah.data.model.api.login
 
-final class LoginRequest {
-}
+//data class LoginRequest()
