+++ android/app/src/main/java/com/littlekakao/travelblahblah/data/model/api/login/LoginResponse.kt
@@ -1,4 +1,3 @@
 package com.littlekakao.travelblahblah.data.model.api.login
 
-final class LoginResponse {
-}
+//data class LoginResponse()
