+++ android/app/src/main/java/com/littlekakao/travelblahblah/data/model/api/login/LogoutResponse.kt
@@ -1,4 +1,3 @@
 package com.littlekakao.travelblahblah.data.model.api.login
 
-class LogoutResponse {
-}
+//data class LogoutResponse()
