+++ android/app/src/main/java/com/littlekakao/travelblahblah/data/repository/ScheduleRepository.kt
@@ -1,0 +1,4 @@
+package com.littlekakao.travelblahblah.data.repository
+
+interface ScheduleRepository {
+}
